+++ app/controllers/attachments_controller.rb
@@ -5,6 +5,7 @@ class AttachmentsController < ApplicationController
 
   def destroy
     @attachment = Attachment.find(params[:id])
+    authorize @attachment
     respond_with @attachment.destroy if @attachment.attachable.user_id == current_user.id
   end
 end
